+++ android/src/main/kotlin/com/adyen/checkout/flutter/components/instant/advanced/IdealComponentAdvancedCallback.kt
@@ -1,0 +1,37 @@
+package com.adyen.checkout.flutter.components.instant.advanced
+
+import ComponentCommunicationModel
+import ComponentFlutterInterface
+import com.adyen.checkout.components.core.ComponentError
+import com.adyen.checkout.components.core.PaymentComponentData
+import com.adyen.checkout.flutter.components.base.ComponentAdvancedCallback
+import com.adyen.checkout.flutter.utils.Constants
+import com.adyen.checkout.ideal.IdealComponentState
+import org.json.JSONObject
+
+class IdealComponentAdvancedCallback(
+    private val componentFlutterApi: ComponentFlutterInterface,
+    private val componentId: String,
+    private val hideLoadingBottomSheet: () -> Unit,
+) : ComponentAdvancedCallback<IdealComponentState>(componentFlutterApi, componentId) {
+    override fun onSubmit(state: IdealComponentState) {
+        val data = PaymentComponentData.SERIALIZER.serialize(state.data)
+        val submitData =
+            JSONObject().apply {
+                put(Constants.ADVANCED_PAYMENT_DATA_KEY, data)
+                put(Constants.ADVANCED_EXTRA_DATA_KEY, null)
+            }
+        val model =
+            ComponentCommunicationModel(
+                ComponentCommunicationType.ONSUBMIT,
+                componentId = componentId,
+                data = submitData.toString(),
+            )
+        componentFlutterApi.onComponentCommunication(model) {}
+    }
+
+    override fun onError(componentError: ComponentError) {
+        hideLoadingBottomSheet()
+        super.onError(componentError)
+    }
+}