+++ android/src/main/kotlin/com/adyen/checkout/flutter/utils/Constants.kt
@@ -4,6 +4,7 @@ class Constants {
     companion object {
         const val WRONG_FLUTTER_ACTIVITY_USAGE_ERROR_MESSAGE =
             "FlutterFragmentActivity not used. Your activity needs to inherit from FlutterFragmentActivity."
+        const val UNKNOWN_PAYMENT_METHOD_TYPE_ERROR_MESSAGE = "Unknown payment method type."
 
         const val GOOGLE_PAY_SESSION_COMPONENT_KEY = "GOOGLE_PAY_SESSION_COMPONENT"
         const val GOOGLE_PAY_ADVANCED_COMPONENT_KEY = "GOOGLE_PAY_ADVANCED_COMPONENT"