+++ example/lib/screens/component/instant/instant_advanced_component_screen.dart
@@ -53,6 +53,8 @@ class InstantAdvancedComponentScreen extends StatelessWidget {
               _extractPaymentMethod(snapshot.data!, "paypal");
           final klarnaPaymentMethodResponse =
               _extractPaymentMethod(snapshot.data!, "klarna");
+          final idealPaymentMethodResponse =
+          _extractPaymentMethod(snapshot.data!, "ideal");
 
           return Column(
             mainAxisAlignment: MainAxisAlignment.center,
@@ -93,7 +95,23 @@ class InstantAdvancedComponentScreen extends StatelessWidget {
                       }
                     });
                   },
-                  child: const Text("Klarna"))
+                  child: const Text("Klarna")),
+              TextButton(
+                  onPressed: () {
+                    AdyenCheckout.advanced
+                        .startInstantComponent(
+                      configuration: instantComponentConfiguration,
+                      paymentMethod: idealPaymentMethodResponse,
+                      checkout: advancedCheckout,
+                    )
+                        .then((paymentResult) {
+                      if (context.mounted) {
+                        DialogBuilder.showPaymentResultDialog(
+                            paymentResult, context);
+                      }
+                    });
+                  },
+                  child: const Text("iDEAL"))
             ],
           );
         } else {
@@ -107,7 +125,7 @@ class InstantAdvancedComponentScreen extends StatelessWidget {
       Map<String, dynamic> paymentMethods, String key) {
     return paymentMethods["paymentMethods"].firstWhere(
       (paymentMethod) => paymentMethod["type"] == key,
-      orElse: () => throw Exception("$key payment method not provided"),
+      orElse: () => <String,dynamic>{},
     );
   }
 }